app/requirements.txt
pyrogram
sqlalchemy
python-dotenv